basicagent/requirements.txt
# Core dependencies
python-dotenv==1.0.1

# LangChain and LangGraph
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-checkpoint-mongodb==0.1.0
langgraph-sdk==0.1.48
langsmith==0.2.10

# FastAPI and server
fastapi==0.115.6
uvicorn==0.34.0
starlette==0.41.3

# MongoDB
pymongo==4.9.2
motor==3.6.0

# HTTP clients
httpx==0.27.2
httpcore==1.0.7
requests==2.32.3
requests-toolbelt==1.0.0

# GreyNoise API
greynoise==2.2.0

# Data handling
pydantic==2.10.4
pydantic_core==2.27.2
PyYAML==6.0.2
jsonpatch==1.33
jsonpointer==3.0.0

# Utilities
tenacity==9.0.0
packaging==24.2
click==8.1.8
grandalf==0.8

# Network and async
aiohttp==3.11.11
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
frozenlist==1.5.0
multidict==6.1.0
propcache==0.2.1
yarl==1.18.3
anyio==4.7.0
sniffio==1.3.1

# Database
SQLAlchemy==2.0.36
dnspython==2.7.0

# Data types
annotated-types==0.7.0
attrs==24.3.0
msgpack==1.1.0
orjson==3.10.13
numpy==2.2.1

# HTTP and security
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
h11==0.14.0

# Type checking
typing_extensions==4.12.2
